# roads_admin/src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Админка дороги
  description: Сервис для распознование дорожных знаков.
  version: 1.0.1
paths:
...
  /users:
    get:
      description: Получение пользователей
      operationId: read_users
      responses:
        "200":
          description: OK!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    post:
      description: Создание пользователя
      operationId: create_user
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                login:
                  type: string
                password:
                  type: string
                privilege_id:
                  type: integer
      responses:
        "201":
          description: OK!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      description: Обновление пользователя
      operationId: update_user
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                login:
                  type: string
                password:
                  type: string
                privilege_id:
                  type: integer
                is_active:
                  type: boolean
      responses:
        "200":
          description: OK!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Запись не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"

    delete:
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      description: Удаление пользователя
      operationId: delete_user
      responses:
        "204":
          description: OK!
          content:
            text/html; charset=utf-8:
              schema:
                nullable: true
        "404":
          description: Запись не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
  /user:
    get:
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      description: Получение записи по условию
      operationId: get_user
      responses:
        "200":
          description: OK!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Запись не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          description: Идентификатор пользователя
          type: integer
        name:
          description: ФИО пользователя
          type: string
        login:
          description: Логин пользователя
          type: string
        password:
          description: Пароль пользователя
          type: string
        privilege_id:
          description: Индификатор привелегии
          type: integer
        is_active:
          description: Активность пользователя
          type: boolean
        created_on:
          description: Дата создание пользователя
          type: string
          format: date-time
        updated_on:
          description: Дата обновление пользователя
          type: string
          format: date-time
...
    Error404:
      type: object
      properties:
        error:
          description: Ошибка
          type: string
...